♻️(serializers) split serializers to dedicated files
As our serializers.py becomes big, classes are splitted like our models.
