+++ src/backend/marsha/core/serializers/__init__.py
@@ -1,0 +1,8 @@
+"""Make all serializers available from marsha.core.serializers."""
+# pylint: disable=wildcard-import,unused-wildcard-import
+from .account import *  # noqa isort:skip
+from .base import *  # noqa isort:skip
+from .file import *  # noqa isort:skip
+from .playlist import *  # noqa isort:skip
+from .video import *  # noqa isort:skip
+from .xapi import *  # noqa isort:skip
