+++ src/backend/marsha/core/serializers/video.py
@@ -1,12 +1,8 @@
-"""Define the structure of our API responses with Django Rest Framework serializers."""
+"""Structure of Video related models API responses with Django Rest Framework serializers."""
 from datetime import timedelta
-import re
 from urllib.parse import parse_qs, quote_plus, urlencode, urlparse, urlunparse
-import uuid
 
 from django.conf import settings
-from django.contrib.auth import get_user_model
-from django.core.exceptions import ValidationError
 from django.urls import reverse
 from django.utils import timezone
 from django.utils.text import slugify
@@ -15,145 +11,12 @@ from botocore.signers import CloudFrontSigner
 from rest_framework import serializers
 from rest_framework_simplejwt.models import TokenUser
 
-from .defaults import (
-    ERROR,
-    HARVESTED,
-    IDLE,
-    LIVE_CHOICES,
-    PROCESSING,
-    READY,
-    RUNNING,
-    STATE_CHOICES,
-    STOPPED,
-)
-from .models import (
-    Document,
-    Organization,
-    OrganizationAccess,
-    Playlist,
-    Thumbnail,
-    TimedTextTrack,
-    Video,
-)
-from .models.account import ADMINISTRATOR, INSTRUCTOR, LTI_ROLES
-from .utils import cloudfront_utils, time_utils, xmpp_utils
-
-
-UUID_REGEX = (
-    "[0-9A-Fa-f]{8}-[0-9A-Fa-f]{4}-[0-9A-Fa-f]{4}-[0-9A-Fa-f]{4}-[0-9A-Fa-f]{12}"
-)
-EXTENSION_REGEX = '[^.\\/:*?&"<>|\r\n]+'
-# This regex matches keys in AWS for videos, timed text tracks, thumbail and document
-TIMED_TEXT_EXTENSIONS = "|".join(m[0] for m in TimedTextTrack.MODE_CHOICES)
-KEY_PATTERN = (
-    r"^{uuid:s}/(?P<model_name>video|thumbnail|timedtexttrack|document)/(?P<object_id>"
-    r"{uuid:s})/(?P<stamp>[0-9]{{10}})(_[a-z-]{{2,10}}_({tt_ex}))?"
-    # The extension is captured and is optional. If present and the resource has an extension
-    # attribute we will save it in database.
-    r"(\.(?P<extension>{extension:s}))?$"
-).format(uuid=UUID_REGEX, tt_ex=TIMED_TEXT_EXTENSIONS, extension=EXTENSION_REGEX)
-KEY_REGEX = re.compile(KEY_PATTERN)
-
-
-class OrganizationAccessSerializer(serializers.ModelSerializer):
-    """
-    Organization access serializer.
-
-    Represents the link between a user and an organization, with a role that grants a specific
-    level of access.
-    """
-
-    organization_name = serializers.SerializerMethodField()
-
-    class Meta:
-        """Meta for OrganizationAccessSerializer."""
-
-        model = OrganizationAccess
-        fields = [
-            "organization",
-            "organization_name",
-            "role",
-            "user",
-        ]
-
-    def get_organization_name(self, organization_access):
-        """
-        Get field for Serializer Method.
-
-        Add the organization name on organization accesses directly, to avoid nesting an
-        additional object just for one string.
-        """
-        return organization_access.organization.name
-
-
-class UserSerializer(serializers.ModelSerializer):
-    """
-    Regular user serializer.
-
-    Allowlist fields as user objects contain a lot more information
-    than we'd like to expose on the API.
-    """
-
-    organization_accesses = OrganizationAccessSerializer(many=True)
-
-    class Meta:
-        """Meta for UserSerializer."""
-
-        model = get_user_model()
-        fields = [
-            "date_joined",
-            "email",
-            "first_name",
-            "id",
-            "is_staff",
-            "is_superuser",
-            "last_name",
-            "organization_accesses",
-        ]
-
-
-class TimestampField(serializers.DateTimeField):
-    """A serializer field to serialize/deserialize a datetime to a Unix timestamp."""
-
-    def to_representation(self, value):
-        """Convert a datetime value to a Unix timestamp.
-
-        Parameters
-        ----------
-        value: Type[datetime.datetime]
-            The datetime value to convert
-
-        Returns
-        -------
-        integer or `None`
-            Unix timestamp for the datetime value or `None`
-
-        """
-        return time_utils.to_timestamp(value) if value else None
-
-    def to_internal_value(self, value):
-        """Convert a Unix timestamp value to a timezone aware datetime.
-
-        Parameters
-        ----------
-        value: Type[string|integer]
-            The Unix timestamp to convert
-
-        Returns
-        -------
-        datetime.datetime or `None`
-            datetime instance corresponding to the Unix timestamp or `None`
-
-        Raises
-        ------
-        ValidationError
-            when the value passed in argument is not a valid timestamp
-
-        """
-        try:
-            return super().to_internal_value(time_utils.to_datetime(value))
-        except OverflowError as error:
-            raise ValidationError(error) from error
+from ..defaults import IDLE, LIVE_CHOICES, RUNNING, STOPPED
+from ..models import Thumbnail, TimedTextTrack, Video
+from ..models.account import ADMINISTRATOR, INSTRUCTOR, LTI_ROLES
+from ..utils import cloudfront_utils, time_utils, xmpp_utils
+from .base import TimestampField
+from .playlist import PlaylistLiteSerializer
 
 
 class TimedTextTrackSerializer(serializers.ModelSerializer):
@@ -422,56 +285,13 @@ class ThumbnailSerializer(serializers.ModelSerializer):
         return None
 
 
-class OrganizationSerializer(serializers.ModelSerializer):
-    """Serializer to display a complete Organization resource."""
+class UpdateLiveStateSerializer(serializers.Serializer):
+    """A serializer to validate data submitted on the UpdateLiveState API endpoint."""
 
-    class Meta:
-        """Meta for OrganizationSerializer."""
-
-        model = Organization
-        fields = [
-            "consumer_sites",
-            "created_on",
-            "id",
-            "name",
-            "users",
-        ]
-
-
-class PlaylistSerializer(serializers.ModelSerializer):
-    """Serializer to display a complete Playlist resource."""
-
-    class Meta:
-        """Meta for Playlistserializer."""
-
-        model = Playlist
-        fields = [
-            "consumer_site",
-            "created_by",
-            "duplicated_from",
-            "id",
-            "is_portable_to_playlist",
-            "is_portable_to_consumer_site",
-            "is_public",
-            "lti_id",
-            "organization",
-            "portable_to",
-            "title",
-            "users",
-        ]
-
-    def create(self, validated_data):
-        """Create the Playlist object with the passed data."""
-        return Playlist.objects.create(**validated_data)
-
-
-class PlaylistLiteSerializer(serializers.ModelSerializer):
-    """A serializer to display a Playlist resource."""
-
-    class Meta:  # noqa
-        model = Playlist
-        fields = ("title", "lti_id")
-        read_only_fields = ("title", "lti_id")
+    state = serializers.ChoiceField(
+        tuple(c for c in LIVE_CHOICES if c[0] in (IDLE, RUNNING, STOPPED))
+    )
+    logGroupName = serializers.CharField()
 
 
 class VideoBaseSerializer(serializers.ModelSerializer):
@@ -764,257 +584,4 @@ class VideoSelectLTISerializer(VideoBaseSerializer):
         """
         return self.context["request"].build_absolute_uri(
             reverse("video_lti_view", args=[obj.id])
-        )
-
-
-class UpdateStateSerializer(serializers.Serializer):
-    """A serializer to validate data submitted on the UpdateState API endpoint."""
-
-    key = serializers.RegexField(KEY_REGEX)
-    state = serializers.ChoiceField(
-        tuple(c for c in STATE_CHOICES if c[0] in (PROCESSING, READY, ERROR, HARVESTED))
-    )
-    extraParameters = serializers.DictField()
-
-    def get_key_elements(self):
-        """Use a regex to parse elements from the key."""
-        elements = KEY_REGEX.match(self.validated_data["key"]).groupdict()
-        elements["uploaded_on"] = time_utils.to_datetime(elements["stamp"])
-        return elements
-
-
-class UpdateLiveStateSerializer(serializers.Serializer):
-    """A serializer to validate data submitted on the UpdateLiveState API endpoint."""
-
-    state = serializers.ChoiceField(
-        tuple(c for c in LIVE_CHOICES if c[0] in (IDLE, RUNNING, STOPPED))
-    )
-    logGroupName = serializers.CharField()
-
-
-class InitiateUploadSerializer(serializers.Serializer):
-    """A serializer to validate data submitted on the initiate-upload API endoint."""
-
-    filename = serializers.CharField()
-    mimetype = serializers.CharField(allow_blank=True)
-
-
-class VerbSerializer(serializers.Serializer):
-    """Validate the verb in a xAPI statement."""
-
-    id = serializers.URLField()
-
-    display = serializers.DictField()
-
-
-class ExtensionSerializer(serializers.Serializer):
-    """Validate the context in a xAPI statement."""
-
-    extensions = serializers.DictField()
-
-
-class XAPIStatementSerializer(serializers.Serializer):
-    """A serializer to validate a xAPI statement."""
-
-    verb = VerbSerializer()
-    context = ExtensionSerializer()
-    result = ExtensionSerializer(required=False)
-
-    id = serializers.RegexField(
-        re.compile("^{uuid}$".format(uuid=UUID_REGEX)),
-        required=False,
-        default=str(uuid.uuid4()),
-    )
-
-    def validate(self, attrs):
-        """Check if there is no extra arguments in the submitted payload."""
-        unknown_keys = set(self.initial_data.keys()) - set(self.fields.keys())
-        if unknown_keys:
-            raise ValidationError("Got unknown fields: {}".format(unknown_keys))
-        return attrs
-
-
-class DocumentSerializer(serializers.ModelSerializer):
-    """A serializer to display a Document resource."""
-
-    class Meta:  # noqa
-        model = Document
-        fields = (
-            "active_stamp",
-            "extension",
-            "filename",
-            "id",
-            "is_ready_to_show",
-            "title",
-            "upload_state",
-            "url",
-            "show_download",
-            "playlist",
-        )
-        read_only_fields = (
-            "active_stamp",
-            "extension",
-            "filename",
-            "id",
-            "is_ready_to_show",
-            "upload_state",
-            "url",
-        )
-
-    active_stamp = TimestampField(
-        source="uploaded_on", required=False, allow_null=True, read_only=True
-    )
-    filename = serializers.SerializerMethodField()
-    url = serializers.SerializerMethodField()
-    is_ready_to_show = serializers.BooleanField(read_only=True)
-    playlist = PlaylistLiteSerializer(read_only=True)
-
-    def _get_extension_string(self, obj):
-        """Document extension with the leading dot.
-
-        Parameters
-        ----------
-        obj : Type[models.Document]
-            The document that we want to serialize
-
-        Returns
-        -------
-        String
-            The document with the leading dot if the document has an extension
-            An empty string otherwise
-
-        """
-        return "." + obj.extension if obj.extension else ""
-
-    def get_filename(self, obj):
-        """Filename of the Document.
-
-        Parameters
-        ----------
-        obj : Type[models.Document]
-            The document that we want to serialize
-
-        Returns
-        -------
-        String
-            The document's filename
-
-        """
-        return "{playlist_title:s}_{title:s}{extension:s}".format(
-            playlist_title=slugify(obj.playlist.title),
-            title=slugify(obj.title),
-            extension=self._get_extension_string(obj),
-        )
-
-    def get_url(self, obj):
-        """Url of the Document.
-
-        Parameters
-        ----------
-        obj : Type[models.Document]
-            The document that we want to serialize
-
-        Returns
-        -------
-        String or None
-            the url to fetch the document on CloudFront
-            None if the document is still not uploaded to S3 with success
-
-        """
-        if obj.uploaded_on is None:
-            return None
-
-        url = (
-            "{protocol:s}://{cloudfront:s}/{pk!s}/document/{stamp:s}{extension:s}"
-            "?response-content-disposition={content_disposition:s}"
-        ).format(
-            protocol=settings.AWS_S3_URL_PROTOCOL,
-            cloudfront=settings.CLOUDFRONT_DOMAIN,
-            pk=obj.pk,
-            stamp=time_utils.to_timestamp(obj.uploaded_on),
-            content_disposition=quote_plus(
-                "attachment; filename=" + self.get_filename(obj)
-            ),
-            extension=self._get_extension_string(obj),
-        )
-
-        # Sign the document urls only if the functionality is activated
-        if settings.CLOUDFRONT_SIGNED_URLS_ACTIVE:
-            date_less_than = timezone.now() + timedelta(
-                seconds=settings.CLOUDFRONT_SIGNED_URLS_VALIDITY
-            )
-            cloudfront_signer = CloudFrontSigner(
-                settings.CLOUDFRONT_ACCESS_KEY_ID, cloudfront_utils.rsa_signer
-            )
-            url = cloudfront_signer.generate_presigned_url(
-                url, date_less_than=date_less_than
-            )
-
-        return url
-
-    def validate_title(self, value):
-        """Force extension removal in the title field (if any).
-
-        Parameters
-        ----------
-        value : Type[string]
-            the value sent in the request
-
-        Returns
-        -------
-        String
-            The title without the extension if there is one.
-
-        """
-        match = re.match(
-            r"^(?P<title>.*)(\.{extension_regex:s})$".format(
-                extension_regex=EXTENSION_REGEX
-            ),
-            value,
-        )
-        if match:
-            return match.group("title")
-
-        return value
-
-
-class DocumentSelectLTISerializer(serializers.ModelSerializer):
-    """A serializer to display a Document resource for LTI select content request."""
-
-    class Meta:  # noqa
-        model = Document
-        fields = (
-            "id",
-            "is_ready_to_show",
-            "title",
-            "upload_state",
-            "lti_url",
-        )
-        read_only_fields = (
-            "id",
-            "is_ready_to_show",
-            "title",
-            "upload_state",
-            "lti_url",
-        )
-
-    is_ready_to_show = serializers.BooleanField(read_only=True)
-    lti_url = serializers.SerializerMethodField()
-
-    def get_lti_url(self, obj):
-        """LTI Url of the Document.
-
-        Parameters
-        ----------
-        obj : Type[models.Document]
-            The document that we want to serialize
-
-        Returns
-        -------
-        String
-            the LTI url to be used by LTI consumers
-
-        """
-        return self.context["request"].build_absolute_uri(
-            reverse("document_lti_view", args=[obj.id])
         )
