+++ src/backend/marsha/core/tests/test_views_lti_respond.py
@@ -1,12 +1,12 @@
 """Test the LTI select view."""
-from unittest import mock
+import random
+from urllib.parse import unquote
 
 from django.test import TestCase
 
-from pylti.common import LTIException
+from oauthlib import oauth1
 
 from ..factories import ConsumerSiteLTIPassportFactory
-from ..lti import LTI
 
 
 # We don't enforce arguments documentation in tests
@@ -18,35 +18,69 @@ class RespondLTIViewTestCase(TestCase):
 
     maxDiff = None
 
-    @mock.patch.object(LTI, "verify")
-    @mock.patch.object(LTI, "get_consumer_site")
-    def test_views_lti_respond(self, mock_get_consumer_site, _mock_verify):
-        """Validate the format of the response returned by the view for an instructor request."""
-        passport = ConsumerSiteLTIPassportFactory()
-
-        # https://www.imsglobal.org/specs/lticiv1p0/specification
-        data = {
+    @staticmethod
+    def _get_signed_lti_parameters():
+        """Generate signed LTI parameters."""
+        passport = ConsumerSiteLTIPassportFactory(consumer_site__domain="testserver")
+        lti_parameters = {
+            "roles": random.choice(["instructor", "administrator"]),
             "content_item_return_url": "http://return.url/",
             "content_items": "some content items",
             "context_id": "unknown",
-            "oauth_consumer_key": passport.oauth_consumer_key,
-            "roles": "Instructor,Administrator",
         }
+        url = "http://testserver/lti/respond/"
+        client = oauth1.Client(
+            client_key=passport.oauth_consumer_key, client_secret=passport.shared_secret
+        )
+        # Compute Authorization header which looks like:
+        # Authorization: OAuth oauth_nonce="80966668944732164491378916897",
+        # oauth_timestamp="1378916897", oauth_version="1.0", oauth_signature_method="HMAC-SHA1",
+        # oauth_consumer_key="", oauth_signature="frVp4JuvT1mVXlxktiAUjQ7%2F1cw%3D"
+        _uri, headers, _body = client.sign(
+            url,
+            http_method="POST",
+            body=lti_parameters,
+            headers={"Content-Type": "application/x-www-form-urlencoded"},
+        )
 
-        mock_get_consumer_site.return_value = passport.consumer_site
+        # Parse headers to pass to template as part of context:
+        oauth_dict = dict(
+            param.strip().replace('"', "").split("=")
+            for param in headers["Authorization"].split(",")
+        )
 
-        response = self.client.post("/lti/respond/", data)
+        signature = oauth_dict["oauth_signature"]
+        oauth_dict["oauth_signature"] = unquote(signature)
+        oauth_dict["oauth_nonce"] = oauth_dict.pop("OAuth oauth_nonce")
+
+        lti_parameters.update(oauth_dict)
+        return lti_parameters
+
+    def test_views_lti_respond(self):
+        """Validate the format of the response returned by the view for an instructor request."""
+        lti_parameters = self._get_signed_lti_parameters()
+        response = self.client.post(
+            "/lti/respond/",
+            lti_parameters,
+            HTTP_REFERER="https://testserver",
+        )
         self.assertEqual(response.status_code, 200)
         self.assertContains(response, "<html>")
 
         context = response.context_data
 
         self.assertEqual(
-            context.get("form_action"), data.get("content_item_return_url")
+            context.get("form_action"),
+            lti_parameters.get("content_item_return_url"),
         )
         form_data = context.get("form_data")
-        for key, value in data.items():
-            if key in ("content_item_return_url", "oauth_signature"):
+        for key, value in lti_parameters.items():
+            if key in (
+                "content_item_return_url",
+                "oauth_signature",
+                "oauth_timestamp",
+                "oauth_nonce",
+            ):
                 continue
             self.assertEqual(value, form_data.get(key))
 
@@ -55,31 +89,33 @@ class RespondLTIViewTestCase(TestCase):
         self.assertContains(response, "oauth_nonce")
 
         self.assertEqual(
-            data.get("oauth_consumer_key"), form_data.get("oauth_consumer_key")
+            lti_parameters.get("oauth_consumer_key"),
+            form_data.get("oauth_consumer_key"),
         )
         self.assertNotEqual(
-            data.get("oauth_signature"), form_data.get("oauth_signature")
+            lti_parameters.get("oauth_signature"), form_data.get("oauth_signature")
         )
 
-    @mock.patch.object(LTI, "verify", side_effect=LTIException)
-    @mock.patch.object(LTI, "get_consumer_site")
-    def test_views_lti_respond_verification_fails(
-        self, mock_get_consumer_site, _mock_verify
-    ):
-        """Validate the format of the response returned by the view for an instructor request."""
-        passport = ConsumerSiteLTIPassportFactory()
+    def test_views_lti_respond_verification_wrong_signature(self):
+        """Wrong signature should raise a 403 error."""
+        lti_parameters = self._get_signed_lti_parameters()
+        lti_parameters["oauth_signature"] = "{:s}a".format(
+            lti_parameters["oauth_signature"]
+        )
 
-        # https://www.imsglobal.org/specs/lticiv1p0/specification
-        data = {
-            "content_item_return_url": "http://return.url/",
-            "content_items": "some content items",
-            "context_id": "unknown",
-            "oauth_consumer_key": passport.oauth_consumer_key,
-            "oauth_signature": "any signature",
-            "roles": "Instructor,Administrator",
-        }
+        response = self.client.post(
+            "/lti/respond/",
+            lti_parameters,
+            HTTP_REFERER="https://testserver",
+        )
+        self.assertEqual(response.status_code, 403)
 
-        mock_get_consumer_site.return_value = passport.consumer_site
-
-        response = self.client.post("/lti/respond/", data)
+    def test_views_lti_respond_verification_wrong_referer(self):
+        """Wrong referer should raise a 403 error."""
+        lti_parameters = self._get_signed_lti_parameters()
+        response = self.client.post(
+            "/lti/respond/",
+            lti_parameters,
+            HTTP_REFERER="https://wrongserver",
+        )
         self.assertEqual(response.status_code, 403)
